New cabal flags approach
We now do flag assignment in the .cabal file ourselves based on a
cabbage.config file that becomes part of the source of the package.

Multiple approaches to freezing.

We now intersect cabal.config files, and can also update version
constraints within a cabal file. The latter is a bit odd, but seemed to
help with some builds.