Added cabbage patching.
This is a way of avoiding conflicts with globally installed packages. We
add a dummy version suffix to the package's sdist (the PLU code for Napa
Cabbage), then adjust all frozen build plans that refer to globally
installed packages to instead refer to the cabbage patched
versions. This lets us rebuild such packages with varying dependency chains.