Big speedup: re-use cached build plans
We now check the /nix/store for cached ~cabal freeze~ output for the
package we're working on when generating the Nix expression. As this was
the slowest part of making a cabbage, everything gets faster.

Note that the first time a particular version of a particular package is
built, Cabal will need to work out the transitive dependencies, but
subsequent builds will not.