Fixed more framework-related bugs.
GHC needs to use frameworks when invoking a C compiler for things like
hsc2hs.

We also have to do a bit of juggling to deal with the way nixpkgs does
things as of today. In particular, the CoreFoundation framework is
broken up into CF and cf-private. Some headers are only available in
cf-private. Having two CoreFoundation.framework directories in the
search path can confuse things, so we put cf-private at the front of the
list of framework dependencies. I am not sure if this will always work,
but it does for the bindings-GLFW package.